Windows ARM64 support (arch/arm64 + backend/arm64 + AArch64 decoder) (#2)
* arm64: Windows ARM64 backend — decoder, backend glue, CMake + CI

Enable inline hooking on Windows ARM64. The arch/arm64 relocator, reader and
writer plus the backend/arm64 interceptor and cpu-context were already present
(extracted from frida-gum) but never compiled: they decode instructions via the
capstone-style hx_disasm API, and hoox only ever shipped an x86 decoder.

- src/disasm/hx_disasm_arm64.c: new compact AArch64 decoder backing the shim.
  Precisely decodes the PC-relative and branch/control instructions the
  relocator/reader/interceptor rewrite or key off (ADR/ADRP, literal LDR/LDRSW,
  B/B.cond, BL, CBZ/CBNZ, TBZ/TBNZ, BR/BLR/RET, SVC) plus MOV(reg) and STP; all
  other instructions are returned valid-but-unclassified so the relocator copies
  them verbatim, while general-purpose register operands are still reported
  conservatively so scratch-register selection never picks a live register.
- src/disasm/hx_disasm.h: add the arm64 type surface (registers, insn ids,
  condition codes, operands, detail) matching the writer's encoding assumptions
  (X0..X30 contiguous, cc == field+1), plus HX_ARCH_AARCH64 / endian modes.
- backend/windows: add hoox_process_get_native_os (the arm64 writer needs it;
  previously only the Linux backend defined it). TLS already falls back to
  TlsGetValue on non-x86, so the Windows backend is otherwise arch-neutral.
- src/CMakeLists.txt: wire arch/arm64 + backend/arm64 + hx_disasm_arm64 and
  define HAVE_ARM64 for the arm64 family.
- tests: golden-vector decode test (arch-neutral — validates the decoder on
  every host, including x86 CI) and a native windows-11-arm build+test CI job.
  Windows 32-bit ARM is marked N/A (ARM64-only platform).


* arm64: fix build — full S/D/Q reg enum, BE byte-order macros, HX_ARCH_ARM64

Caught by compiling the arch/arm64 sources against the new shim:
- hx_disasm.h: enumerate every S0..S31 / D0..D31 / Q0..Q31 register (the writer
  names them individually, e.g. HX_ARM64_REG_Q1); and define HX_ARCH_ARM64
  unconditionally (hooxdefs.h only defines it when HX_API_MAJOR >= 6, which hoox
  does not set).
- hxdefs.h: add the *_TO_BE / *_FROM_BE integer byte-order macros the arm64
  writer uses (only the _LE variants existed — the x86 writer never needed BE).


* arm64: silence MSVC /WX narrowing warnings (C4244/C4267)

MSVC on ARM64 is LLP64 (hx_uint/hx_uint32 are 32-bit, pointers/size_t 64-bit),
so the arm64 writer/relocator and the Windows TLS fallback tripped
warnings-as-errors on int64/size_t -> 32-bit narrowing. Add explicit casts at
each site (all provably in-range: instruction-encoding ORs, guarded 32-bit
literals, TLS key -> DWORD, pointer-diff instruction counts).

Validated locally by cross-compiling every arm64 source with
`clang --target=aarch64-pc-windows-msvc -Wshorten-64-to-32` (MSVC-parity
narrowing diagnostics), which is what surfaced four writer sites the first CI
run did not reach.


* test(memory): arch-aware machine code — fix ARM64 "Illegal instruction"

The M3 memory smoke test hardcoded x86-64 `mov eax,imm; ret` (B8/C3) and
executed it, which faulted on ARM64. Emit the native `mov w0,#imm; ret`
(MOVZ + RET) on arm64 instead. On the windows-11-arm runner the interceptor
behaviour suite and all other tests already pass; this was the last x86-only
assumption in the test tree.


* docs: mark Windows ARM64 supported (ARM32 N/A)

Update the platform-support matrix and status prose in README.md / README_en.md
and CLAUDE.md now that Windows ARM64 builds and passes the full test suite on the
native windows-11-arm CI runner. Windows 32-bit ARM is marked N/A (Windows on
ARM is ARM64-only).


---------

+++ src/backend/windows/hooxtls-windows.c
@@ -265,14 +265,14 @@ hoox_tls_key_set_value (HooxTlsKey key,
 hx_pointer
 hoox_tls_key_get_value (HooxTlsKey key)
 {
-  return TlsGetValue (key);
+  return TlsGetValue ((DWORD) key);
 }
 
 void
 hoox_tls_key_set_value (HooxTlsKey key,
                        hx_pointer value)
 {
-  TlsSetValue (key, value);
+  TlsSetValue ((DWORD) key, value);
 }
 
 #endif
